// src/api/model/modSummary.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * 5-thirty
 * backend data aggregate for 6 o'clock
 * OpenAPI spec version: 0.0.1
 */
import type { UserHead } from "./userHead";

export interface ModSummary {
  claimed: number;
  days: number;
  handled: number;
  head?: UserHead;
  userId: number;
}